Not trying to pull collector:master

+++ kernel-modules/support-packages/02-fetch-collectors-metadata.sh
@@ -16,18 +16,25 @@ for version_dir in "${MD_DIR}/collector-versions"/*; do
     [[ -d "$version_dir" ]] || continue
     version="$(basename "$version_dir")"
 
-    collector_image="quay.io/rhacs-eng/collector:${version}"
-    docker pull "$collector_image"
-    tmp_output="$(mktemp)"
+    if [[ "$version" != master ]]; then
+        collector_image="quay.io/rhacs-eng/collector:${version}"
+        docker pull "$collector_image"
+        tmp_output="$(mktemp)"
 
-    docker run --rm --entrypoint /bin/bash "$collector_image" -c '
-        cd /kernel-modules &&
-        head -n 1 MODULE_VERSION.txt &&
-        ls *.gz
-    ' > "$tmp_output"
+        docker run --rm --entrypoint /bin/bash "$collector_image" -c '
+            cd /kernel-modules &&
+            head -n 1 MODULE_VERSION.txt &&
+            ls *.gz
+        ' > "$tmp_output"
 
-    head -n 1 "$tmp_output" > "${version_dir}/MODULE_VERSION"
-    tail -n +2 "$tmp_output" > "${version_dir}/INVENTORY"
-    rm -f "$tmp_output" || true
-    docker image rm "$collector_image"
+        head -n 1 "$tmp_output" > "${version_dir}/MODULE_VERSION"
+        tail -n +2 "$tmp_output" > "${version_dir}/INVENTORY"
+        rm -f "$tmp_output" || true
+        docker image rm "$collector_image"
+    fi
 done
+
+master_dir="${MD_DIR}/collector-versions/master"
+mkdir -p "${master_dir}"
+git checkout master -- "${SOURCE_ROOT}/kernel-modules/MODULE_VERSION"
+cp "${SOURCE_ROOT}/kernel-modules/MODULE_VERSION" "$master_dir"
